+++ core/api/golang/lib/services/container_config.go
@@ -17,6 +17,10 @@
 
 package services
 
+const (
+	useStaticPrivatePorts = true
+)
+
 // The UUID of an artifact containing files that should be mounted into a service container
 type FilesArtifactUUID string
 
@@ -26,9 +30,10 @@ type FilesArtifactUUID string
 // Docs available at https://docs.kurtosistech.com/kurtosis-core/lib-documentation
 type ContainerConfig struct {
 	image                        string
-	usedPorts                   map[string]*PortSpec
-	filesArtifactMountpoints    map[FilesArtifactUUID]string
-	entrypointOverrideArgs      []string
+	usedPorts                    map[string]*PortSpec
+	useStaticPrivatePorts        bool //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
+	filesArtifactMountpoints     map[FilesArtifactUUID]string
+	entrypointOverrideArgs       []string
 	cmdOverrideArgs              []string
 	environmentVariableOverrides map[string]string
 }
@@ -57,6 +62,11 @@ func (config *ContainerConfig) GetEnvironmentVariableOverrides() map[string]stri
 	return config.environmentVariableOverrides
 }
 
+//TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
+func (config *ContainerConfig) GetUseStaticPrivatePorts() bool {
+	return config.useStaticPrivatePorts
+}
+
 // ====================================================================================================
 //                                      Builder
 // ====================================================================================================
@@ -65,6 +75,7 @@ func (config *ContainerConfig) GetEnvironmentVariableOverrides() map[string]stri
 type ContainerConfigBuilder struct {
 	image                        string
 	usedPorts                    map[string]*PortSpec
+	useStaticPrivatePorts        bool //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
 	filesArtifactMountpoints     map[FilesArtifactUUID]string
 	entrypointOverrideArgs       []string
 	cmdOverrideArgs              []string
@@ -107,6 +118,12 @@ func (builder *ContainerConfigBuilder) WithEnvironmentVariableOverrides(envVars 
 	return builder
 }
 
+//TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
+func (builder *ContainerConfigBuilder) WithUseStaticPrivatePorts() *ContainerConfigBuilder {
+	builder.useStaticPrivatePorts = useStaticPrivatePorts
+	return builder
+}
+
 func (builder *ContainerConfigBuilder) Build() *ContainerConfig {
 	return &ContainerConfig{
 		image:                        builder.image,
@@ -115,5 +132,7 @@ func (builder *ContainerConfigBuilder) Build() *ContainerConfig {
 		entrypointOverrideArgs:       builder.entrypointOverrideArgs,
 		cmdOverrideArgs:              builder.cmdOverrideArgs,
 		environmentVariableOverrides: builder.environmentVariableOverrides,
+		//TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
+		useStaticPrivatePorts: 		  builder.useStaticPrivatePorts,
 	}
 }