+++ core/api/typescript/src/lib/enclaves/enclave_context.ts
@@ -274,6 +274,7 @@ export class EnclaveContext {
             serviceId,
             containerConfig.image,
             privatePortsForApi,
+            containerConfig.useStaticPrivatePorts,
             containerConfig.entrypointOverrideArgs,
             containerConfig.cmdOverrideArgs,
             containerConfig.environmentVariableOverrides,