+++ core/api/typescript/src/lib/services/container_config.ts
@@ -3,6 +3,9 @@ import { PortSpec } from "./port_spec";
 // The UUID of an artifact containing files that should be mounted into a service container
 export type FilesArtifactUUID = string;
 
+const USE_STATIC_PRIVATE_PORTS = true;
+const DO_NOT_USE_STATIC_PRIVATE_PORTS = false;
+
 // ====================================================================================================
 //                                    Config Object
 // ====================================================================================================
@@ -12,6 +15,7 @@ export class ContainerConfig {
     constructor(
         public readonly image: string,
         public readonly usedPorts: Map<string, PortSpec>,
+        public readonly useStaticPrivatePorts: boolean, //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
         public readonly filesArtifactMountpoints: Map<FilesArtifactUUID, string>,
         public readonly entrypointOverrideArgs: string[],
         public readonly cmdOverrideArgs: string[],
@@ -29,6 +33,7 @@ export class ContainerConfig {
 export class ContainerConfigBuilder {
     private readonly image: string;
     private usedPorts: Map<string, PortSpec>;
+    private useStaticPrivatePorts: boolean; //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
     private filesArtifactMountpoints: Map<FilesArtifactUUID, string>;
     private entrypointOverrideArgs: string[];
 	private cmdOverrideArgs: string[];
@@ -41,6 +46,7 @@ export class ContainerConfigBuilder {
         this.entrypointOverrideArgs = [];
         this.cmdOverrideArgs = [];
         this.environmentVariableOverrides = new Map();
+        this.useStaticPrivatePorts = DO_NOT_USE_STATIC_PRIVATE_PORTS; //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
     }
 
     public withUsedPorts(usedPorts: Map<string, PortSpec>): ContainerConfigBuilder {
@@ -68,14 +74,21 @@ export class ContainerConfigBuilder {
         return this;
 	}
 
+    //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
+    public withUseStaticPrivatePorts(): ContainerConfigBuilder {
+        this.useStaticPrivatePorts = USE_STATIC_PRIVATE_PORTS;
+        return this;
+    }
+
     public build(): ContainerConfig {
         return new ContainerConfig(
             this.image,
             this.usedPorts,
+            this.useStaticPrivatePorts,
             this.filesArtifactMountpoints,
             this.entrypointOverrideArgs,
             this.cmdOverrideArgs,
-            this.environmentVariableOverrides
+            this.environmentVariableOverrides,
         );
     }
 }